feat(fetch-releases): increase fetched release limits to 100
Increase the number of releases from GitHub from 10 to 100
and also raise the number of prereleases included from 10 to 100.

This change reduces the chance of missing the "last played" release by
requesting a larger window of releases. Add a TODO to fetch the last
played release explicitly in the future if a more efficient approach is
desired.